refactor: Move easing utilities to aux-common
